# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Raqib Api",
        "description": "API documentation for My App",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/v1/business-mission": {
            "get": {
                "tags": [
                    "Business Missions"
                ],
                "summary": "Get list of business missions",
                "operationId": "f15072c82d6d070042be20604274c538",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BusinessMission"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Permission denied"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/business-mission/store": {
            "post": {
                "tags": [
                    "Business Missions"
                ],
                "summary": "Create a new business mission",
                "operationId": "e7d3465c4bc1d8798b948464f0900ee2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "employee_id",
                                    "title",
                                    "start_date",
                                    "end_date",
                                    "description"
                                ],
                                "properties": {
                                    "employee_id": {
                                        "type": "integer"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "remark": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Business mission created"
                    },
                    "403": {
                        "description": "Permission denied"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/business-mission/{id}": {
            "get": {
                "tags": [
                    "Business Missions"
                ],
                "summary": "Get a specific business mission",
                "operationId": "baf0c838c307ec3f80fc7a1b983aff3a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Business mission ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Business Missions"
                ],
                "summary": "Update a business mission",
                "operationId": "3fa9d38fb030b87b0d0200c12cbaef18",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Business mission ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "employee_id",
                                    "title",
                                    "start_date",
                                    "end_date",
                                    "description"
                                ],
                                "properties": {
                                    "employee_id": {
                                        "type": "integer"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "remark": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "Pending",
                                            "Approved",
                                            "In_Progress",
                                            "Completed",
                                            "Canceled"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Business mission updated"
                    },
                    "403": {
                        "description": "Permission denied"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Business Missions"
                ],
                "summary": "Delete a business mission",
                "operationId": "1efa8cdbf72aee3508ce8fed43ce4f4e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Business mission ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted successfully"
                    },
                    "403": {
                        "description": "Permission denied"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/business-mission/{id}/reject": {
            "put": {
                "tags": [
                    "Business Missions"
                ],
                "summary": "Reject a business mission",
                "operationId": "bbeb314fd625374bf7468beb7ea18c88",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Business mission approval ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "action": {
                                        "type": "string",
                                        "enum": [
                                            "reject"
                                        ],
                                        "example": "reject"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rejection successful"
                    },
                    "403": {
                        "description": "Permission denied"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/business-mission/pending-approvals": {
            "get": {
                "tags": [
                    "Business Missions"
                ],
                "summary": "Get pending business mission approvals for current user",
                "operationId": "f39a560d39d2975726d16f03505a14f7",
                "responses": {
                    "200": {
                        "description": "List of pending approvals"
                    },
                    "403": {
                        "description": "Permission denied"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "BusinessMission": {
                "title": "Business Mission",
                "description": "Details about a business mission",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "employee_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "title": {
                        "type": "string",
                        "example": "Client Meeting in Dubai"
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-07-20"
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-07-25"
                    },
                    "description": {
                        "type": "text",
                        "example": "Meeting with key client to discuss project status."
                    },
                    "remark": {
                        "type": "string",
                        "example": "Don't forget the presentation slides.",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "Pending",
                            "Approved",
                            "In_Progress",
                            "Completed",
                            "Canceled",
                            "Rejected"
                        ],
                        "example": "Pending"
                    },
                    "employee": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 12
                            },
                            "name": {
                                "type": "string",
                                "example": "John Doe"
                            }
                        },
                        "type": "object"
                    },
                    "approvals": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "approver_id": {
                                    "type": "integer",
                                    "example": 8
                                },
                                "status": {
                                    "type": "string",
                                    "example": "pending"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}